Rename AppContext to OnboardingContext, integrate use-onboarding-sections hook, move relevant files to onboarding folder
