+++ src/loading-screen/index.ts
@@ -1,2 +1,3 @@
 export {LoadingScreenBoundary} from './LoadingScreenBoundary';
 export {LoadingScreen} from './LoadingScreen';
+export {useIsLoadingAppState} from './use-is-loading-app-state';
