+++ src/loading-screen/use-is-loading-app-state.ts
@@ -1,0 +1,6 @@
+import {useOnboardingState} from '@atb/onboarding';
+
+export const useIsLoadingAppState = () => {
+  const {isLoading: onboardingIsLoading} = useOnboardingState();
+  return onboardingIsLoading; // also wait for e.g. popover flags ++?
+};
