+++ src/onboarding/index.ts
@@ -4,3 +4,15 @@ export {
 } from './OnboardingContext';
 export {OnboardingScreenComponent} from './OnboardingScreenComponent';
 export {OnboardingFullScreenView} from './OnboardingFullScreenView';
+export {
+  useOnboardingSectionIsOnboarded,
+  getOnboardingSectionIsOnboarded,
+} from './use-onboarding-section-is-onboarded';
+export {useOnboardingGetCustomShouldShow} from './use-onboarding-get-custom-should-show';
+export {staticOnboardingSectionsInPrioritizedOrder} from './onboarding-config';
+export type {
+  OnboardingSectionId,
+  OnboardingSection,
+  LoadedOnboardingSection,
+  StaticOnboardingSection,
+} from './types';
