+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_DebugInfoScreen.tsx
@@ -58,7 +58,6 @@ import {useBeaconsState} from '@atb/beacons/BeaconsContext';
 import {useOnBehalfOfEnabledDebugOverride} from '@atb/on-behalf-of';
 import {useTicketInformationEnabledDebugOverride} from '@atb/stacks-hierarchy/Root_PurchaseOverviewScreen/use-is-ticket-information-enabled';
 import {usePosthogEnabledDebugOverride} from '@atb/analytics/use-is-posthog-enabled';
-import {useOnboardingSections} from '@atb/utils/use-onboarding-sections';
 import {useOnboardingState} from '@atb/onboarding';
 
 function setClipboard(content: string) {
@@ -69,9 +68,11 @@ function setClipboard(content: string) {
 export const Profile_DebugInfoScreen = () => {
   const style = useProfileHomeStyle();
 
-  const {restartOnboardingSection, restartAllOnboardingSections} =
-    useOnboardingState();
-  const onboardingSections = useOnboardingSections(false);
+  const {
+    onboardingSections,
+    restartOnboardingSection,
+    restartAllOnboardingSections,
+  } = useOnboardingState();
 
   const {
     onboardForBeacons,
